[compiler] Inverse degenerate sub-group cloning and offer opt-out
Some sub-group builtins have no work-group equivalents (the incoming
shuffles). In this situation, we are unable to offer degenerate
sub-groups as an option.

To accomodate this, we need to offer an opt-out clause in the degenerate
sub-groups pass. This is done first by 'poisoning' the call-graph of
sub-group users, propagating impossible sub-group builtins up through
all their users.

Then we invert the cloning process to leave the original functions as
the non-degenerate sub-group functions where possible, with the cloned
ones becoming the degenerate sub-groups. This is the opposite of what it
was before. This is so that we can easily leave functions in their place
if they don't need updating. With the old process we'd globally rename
the existing functions to their degenerate equivalents:

    `declare void @foo()` -> `declare void @foo.degenerate-subgroups`

This would pre-emptively update a non-degenerate user `@bar` which calls
`@foo` to calling the degenerate function `@foo.degenerate-subgroups`.
Then we'd have to go back over `@bar` and patch in the correct call to
`@foo`.

It's far easier, if we're changing some functions & altering their
behaviour while leaving some untouched, to do just that.

This alters the order of functions coming out of this pass (degenerate
ones are now found at the end), but otherwise everything looks okay.