[compiler] Encode 'no' and 'degenerate' subgroups distinctly
This commit better conceptualizes that a function's use of degenerate
sub-groups is different from using no sub-groups at all.

The `DegenerateSubGroupPass` would confuse the two, marking functions
that it couldn't detect sub-group usage in as using degenerate
sub-groups. As a knock-on effect, the `GenericMetadataAnalysis` would do
this too and encode both concepts as '0'. This effectively led to the
same situation where a kernel using degenerate sub-groups and one using
no sub-groups were legal for any work-group size.

Aside from the confusion of two separate concepts, the problem comes
when the degenerate sub-groups pass is not run by a target. In this
situation, if that target runs the vectorizer, then a sub-group that
doesn't use sub-groups is marked as having a sub-group size equal to the
vectorization factor, which means the runtime must infer that it is
unsafe to run if the work-group size is not a multiple of the
vectorization factor.

This commit changes things so that the two concepts are tracked
separately in the compiler. They come together to mean the same thing in
the ELF metadata (for now - this could be changed in the future) but
there is now no reliance on the `DegenerateSubGroupPass` to set
attributes.

This is done by 'caching' the result of the `SubgroupAnalysis` before
the `DegenerateSubGroupPass` runs as a function attribute. It's
generally assumed that later passes won't introduce any new uses of
sub-groups and, as such, the attribute is held as the truth, including
by future runs of the analysis. If a pass ever did introduce new
expilcit uses of sub-groups to a function, it must remove that attribute
(if present).

This is done early on in the 'generic' pass pipelines, to avoid each
target having to do its own work, and also so that the effects of
inlining do not ruin the analysis (the sub-group analysis can't
determine the usage/non-usage of sub-groups after the mux builtins are
defined and inlined, as `host` does before the `AddMetadataPass` runs).