[mux][compiler] Encode 'no sub-groups' as 1 rather than 0
This undoes a recent change to how 'no sub-groups' are represented in
the ELF metadata. The value 'zero' is once again only for degenerate
sub-groups. No explicit use of sub-groups is instead encoded at 1 (well,
the mux sub-group size but for now that's always 1). This is to
faciliate the existing mux queries which *do* need to distinguish
between 'degenerate' and 'no' sub-groups for API queries.

It also updates the WorkItemLoopsPass to correctly represent the base
sub-group size as the mux sub-group size, rather than assuming 1
universally.